backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.1
email-validator==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
Pillow==10.1.0